fix: don't render blank lines at the bottom
It turns out that the height of the doc window is the number of lines
in the text. The problem here is that each code blocks contain two extra
lines which will be removed by `stylize_markdown` function and thus the
extra space.

The problem does not seem to appear in any of the LSP floating window
because of the order the functions are being invoked. `stylize_markdown`
needs to be invoked first which will remove those code fences and apply
the syntax highlighting and then we should be calculating the height of
the floating window.