Fix regression with showing empty resolve window
I actually fixed this in the past already here:
https://github.com/hrsh7th/nvim-compe/commit/4df0108195dfbc36f6143a7d0e1f6bf6418432b8,

but it looks like is snuck back in. Some servers will still return a
`detail` field in the completion item with an empty string. Currently
this is causing an extra empty window to appear. This small change just
makes sure the detail isn't empty, and if it is, don't show the window.