fix: don't render blank lines at the bottom (#458)
* fix: don't render blank lines at the bottom

It turns out that the height of the doc window is the number of lines
in the text. The problem here is that each code blocks contain two extra
lines which will be removed by `stylize_markdown` function and thus the
extra space.

The problem does not seem to appear in any of the LSP floating window
because of the order the functions are being invoked. `stylize_markdown`
needs to be invoked first which will remove those code fences and apply
the syntax highlighting and then we should be calculating the height of
the floating window.

* fix: allow scroll only upto win height

* fix: return if there is no space available for docs

This was already present previously but I removed it unknowingly. I
started getting this error:

Error executing vim.schedule lua callback: ...nvim/site/pack/packer/opt/nvim-compe/lua/compe/float.lua:111: attempt to index local 'float_options' (a nil value)

* refactor: remove unused lines and shift comment